+++ packages/frontend/src/pages/QuantifyPeriodReceiver/components/DuplicateDialog.tsx
@@ -3,6 +3,9 @@ import { faCalculator, faTimes } from '@fortawesome/free-solid-svg-icons';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { PraiseDto } from 'api/dist/praise/types';
 import PraiseAutosuggest from './PraiseAutosuggest';
+import { PeriodPageParams } from '@/model/periods';
+import { useParams } from 'react-router-dom';
+import { usePeriodSettingValueRealized } from '@/model/periodsettings';
 
 interface DuplicateDialogProps {
   onClose(): void;
@@ -17,6 +20,13 @@ const DuplicateDialog = ({
   open = false,
   praise,
 }: DuplicateDialogProps): JSX.Element | null => {
+  const { periodId } = useParams<PeriodPageParams>();
+
+  const duplicatePraisePercentage = usePeriodSettingValueRealized(
+    periodId,
+    'PRAISE_QUANTIFY_DUPLICATE_PRAISE_PERCENTAGE'
+  ) as number;
+
   if (!praise) return null;
 
   return (
@@ -34,6 +44,12 @@ const DuplicateDialog = ({
           <h2 className="text-center mb-7">
             Mark praise #{praise._id.slice(-4)} as duplicate
           </h2>
+          {duplicatePraisePercentage && (
+            <p className="text-center mb-7">
+              Duplicate praise are given a score that is{' '}
+              {duplicatePraisePercentage * 100}% of the original quantification.
+            </p>
+          )}
 
           <div className="flex justify-center">
             <PraiseAutosuggest