+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Modify settings architecture so Periods have their own unique settings. New periods are created with the "global" settings as default values. #138 #116
 - Verify no quantifiers have been assigned in period before assigning
 - Display Period start and end dates in UTC. All others in users' local timezone. #204
+- Duplicate dialog message: "Duplicate praise are given a score that is X% of the original quantification."
 
 ### Fixed